finally it seems working and now we need to test

+++ quisp/modules/QNIC/StationaryQubit/StationaryQubit.h
@@ -30,7 +30,7 @@ class StationaryQubit : public IStationaryQubit {
   bool checkBusy() override;
   void setFree(bool consumed) override;
   /*In use. E.g. waiting for purification result.*/
-  void Lock(unsigned long rs_id, int rule_id, int action_id) override;
+  void Lock(unsigned long rs_id, unsigned long rule_id, int action_id) override;
   void Unlock() override;
   bool isLocked() override;
   void Allocate() override;
